Split CaptureOverlay.cpp into focused TUs to manage 5000-line file
